Added lists of tools...
Added lists of tools to Guidebook. Add a Filemaker file to the
repository. The file keeps track of my tools. Added a tab2text.pl which
formats the output of my Filemaker data into plain text. Added a very
rudimentary PDF document -- a version of the Guidebook desgned for
human consumption. For a good time, added an RDF output of my Zotero
database.

"Librarians love lists."